Add some basic level of config validation
This does some strange stuff as you can see from the comments but
gets us some basic validation. Important to note we look at the JSON
after it has been marshalled into a struct so they will always have
default values in them that will need to be accounted for in the
program.

This could be cleaned up if we switch to using pure JSON for config
however I hate that.