Fix possible race condition
This fixes a possible issue where slices modified in the Config struct
were accessed in one go routine while being changed in another. I never
saw this occure and it would be be very rare but it was possible for
this to cause a race before. I think this is the only location in the
program this was possible but no promises.