Remove shell.nix file
Nix is pretty cool in theory but in practice I have just found it to be
a pain in the ass with an unmanageable config language. Possibly using
it outside of NixOS is just asking for pain but at least in the case of
golang it's way easier to use docker or a system install of go.